feat : Extra features added

+++ OrderItem.java
@@ -2,9 +2,9 @@ public class OrderItem{
     Product product;
     int quantity;
     String offer;
-    float totalPrice;
+    int totalPrice;
 
-    OrderItem(Product product, int quantity, String offer, float totalPrice){
+    OrderItem(Product product, int quantity, String offer, int totalPrice){
         this.product = product;
         this.quantity = quantity;
         this.offer = offer;
